layers: Consolidate thread-safety header files
Combine thread-safety.h and thread_safety_validation.h to allow for
more code-gen.
